Fixes #19
Templates in index.html were using Liquid, so I started porting some of
the commands over to Jinja2 instead. This is a start on that, but I had
to remove interactive-table.html from being included in the index
because there are too many commands in there that depend on Liquid.

Still, you can run app.py without errors for the index page now.